Issue #950: extract tau-extensions crate
