+++ crates/tau-coding-agent/Cargo.toml
@@ -22,6 +22,7 @@ tau-gateway = { path = "../tau-gateway" }
 tau-multi-channel = { path = "../tau-multi-channel" }
 tau-session = { path = "../tau-session" }
 tau-skills = { path = "../tau-skills" }
+tau-extensions = { path = "../tau-extensions" }
 tau-orchestrator = { path = "../tau-orchestrator" }
 tau-runtime = { path = "../tau-runtime" }
 tau-agent-core = { path = "../tau-agent-core" }
