+++ crates/tau-extensions/Cargo.toml
@@ -1,0 +1,13 @@
+[package]
+name = "tau-extensions"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+anyhow.workspace = true
+serde.workspace = true
+serde_json.workspace = true
+wait-timeout.workspace = true
+
+[dev-dependencies]
+tempfile = "3"
